feat: changed architecture details
