+++ internal/api/http/http_test.go
@@ -10,116 +10,6 @@ import (
 	"testing"
 )
 
-func TestParseGitURL(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		path    string
-		query   string
-		wantNil bool
-		want    *parsedURL
-	}{
-		{
-			name:  "InfoRefs upload-pack",
-			path:  "/myrepo.git/info/refs",
-			query: "service=git-upload-pack",
-			want: &parsedURL{
-				Repo:    "myrepo.git",
-				Service: infoRefs,
-				Query:   "service=git-upload-pack",
-			},
-		},
-		{
-			name:  "InfoRefs receive-pack",
-			path:  "/myrepo.git/info/refs",
-			query: "service=git-receive-pack",
-			want: &parsedURL{
-				Repo:    "myrepo.git",
-				Service: infoRefs,
-				Query:   "service=git-receive-pack",
-			},
-		},
-		{
-			name:  "UploadPack POST",
-			path:  "/myrepo.git/git-upload-pack",
-			query: "",
-			want: &parsedURL{
-				Repo:    "myrepo.git",
-				Service: uploadPack,
-			},
-		},
-		{
-			name:  "ReceivePack POST",
-			path:  "/myrepo.git/git-receive-pack",
-			query: "",
-			want: &parsedURL{
-				Repo:    "myrepo.git",
-				Service: receivePack,
-			},
-		},
-		{
-			name:  "Nested repo path",
-			path:  "/org/team/repo.git/git-upload-pack",
-			query: "",
-			want: &parsedURL{
-				Repo:    "org/team/repo.git",
-				Service: uploadPack,
-			},
-		},
-		{
-			name:    "Root path",
-			path:    "/",
-			query:   "",
-			wantNil: true,
-		},
-		{
-			name:    "InfoRefs without service query",
-			path:    "/repo/info/refs",
-			query:   "",
-			wantNil: true,
-		},
-		{
-			name:    "Unknown path",
-			path:    "/repo/objects/info/packs",
-			query:   "",
-			wantNil: true,
-		},
-		{
-			name:    "Empty repo in InfoRefs",
-			path:    "/info/refs",
-			query:   "service=git-upload-pack",
-			wantNil: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			p := parseGitURL(tc.path, tc.query)
-			if tc.wantNil {
-				if p != nil {
-					t.Errorf("parseGitURL(%q, %q) = %+v, want nil", tc.path, tc.query, p)
-				}
-				return
-			}
-			if p == nil {
-				t.Fatalf("parseGitURL(%q, %q) = nil, want %+v", tc.path, tc.query, tc.want)
-			}
-			if p.Repo != tc.want.Repo {
-				t.Errorf("repo = %q, want %q", p.Repo, tc.want.Repo)
-			}
-			if p.Service != tc.want.Service {
-				t.Errorf("service = %v, want %v", p.Service, tc.want.Service)
-			}
-			if p.Query != tc.want.Query {
-				t.Errorf("query = %q, want %q", p.Query, tc.want.Query)
-			}
-		})
-	}
-}
-
 func TestServeHTTP_InfoRefs(t *testing.T) {
 	t.Parallel()
 
@@ -128,7 +18,6 @@ func TestServeHTTP_InfoRefs(t *testing.T) {
 	if err := os.MkdirAll(repoPath, 0755); err != nil {
 		t.Fatal(err)
 	}
-	// Initialize a bare git repository
 	if err := runGit(repoPath, "init", "--bare"); err != nil {
 		t.Skipf("git not available: %v", err)
 	}
@@ -157,10 +46,10 @@ func TestServeHTTP_InfoRefs(t *testing.T) {
 			wantContentType: "application/x-git-receive-pack-advertisement",
 		},
 		{
-			name:       "POST to info/refs not allowed",
+			name:       "POST to info/refs not found",
 			url:        "/test-repo.git/info/refs?service=git-upload-pack",
 			method:     http.MethodPost,
-			wantStatus: http.StatusMethodNotAllowed,
+			wantStatus: http.StatusNotFound,
 		},
 		{
 			name:       "GET info/refs without service query",
@@ -218,15 +107,15 @@ func TestServeHTTP_UploadPack(t *testing.T) {
 			name:        "POST git-upload-pack with valid request",
 			url:         "/test-repo.git/git-upload-pack",
 			method:      http.MethodPost,
-			body:        "0000", // flush packet
+			body:        "0000",
 			contentType: "application/x-git-upload-pack-request",
 			wantStatus:  http.StatusOK,
 		},
 		{
-			name:       "GET to git-upload-pack not allowed",
+			name:       "GET to git-upload-pack not found",
 			url:        "/test-repo.git/git-upload-pack",
 			method:     http.MethodGet,
-			wantStatus: http.StatusMethodNotAllowed,
+			wantStatus: http.StatusNotFound,
 		},
 	}
 
@@ -275,15 +164,15 @@ func TestServeHTTP_ReceivePack(t *testing.T) {
 			name:        "POST git-receive-pack with valid request",
 			url:         "/test-repo.git/git-receive-pack",
 			method:      http.MethodPost,
-			body:        "0000", // flush packet
+			body:        "0000",
 			contentType: "application/x-git-receive-pack-request",
 			wantStatus:  http.StatusOK,
 		},
 		{
-			name:       "GET to git-receive-pack not allowed",
+			name:       "GET to git-receive-pack not found",
 			url:        "/test-repo.git/git-receive-pack",
 			method:     http.MethodGet,
-			wantStatus: http.StatusMethodNotAllowed,
+			wantStatus: http.StatusNotFound,
 		},
 	}
 
@@ -316,7 +205,6 @@ func TestServeHTTP_ErrorCases(t *testing.T) {
 		name       string
 		url        string
 		method     string
-		body       string
 		wantStatus int
 	}{
 		{
@@ -343,7 +231,7 @@ func TestServeHTTP_ErrorCases(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
-			req := httptest.NewRequest(tc.method, tc.url, strings.NewReader(tc.body))
+			req := httptest.NewRequest(tc.method, tc.url, nil)
 			rec := httptest.NewRecorder()
 
 			srv.ServeHTTP(rec, req)
@@ -355,11 +243,8 @@ func TestServeHTTP_ErrorCases(t *testing.T) {
 	}
 }
 
-// runGit is a helper to run git commands for test setup.
 func runGit(dir string, args ...string) error {
 	cmd := exec.Command("git", args...)
 	cmd.Dir = dir
-	cmd.Stdout = nil
-	cmd.Stderr = nil
 	return cmd.Run()
 }
