+++ internal/modules/user/repository.go
@@ -1,0 +1,17 @@
+package user
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+)
+
+type Repository interface {
+	Create(ctx context.Context, u User) (User, error)
+	GetByID(ctx context.Context, id uuid.UUID) (User, error)
+	GetByEmail(ctx context.Context, email string) (User, error)
+	GetByUsername(ctx context.Context, username string) (User, error)
+	Verify(ctx context.Context, id uuid.UUID) error
+	SoftDelete(ctx context.Context, id uuid.UUID) error
+	UpdatePassword(ctx context.Context, id uuid.UUID, password string) error
+}
